snapshot_data_ingestion_pipeline/python/requirements.txt
apache-airflow[gcp]
google-cloud-bigquery
google-cloud-pubsub

